requirements.txt
APScheduler==3.6.3
asgiref==3.7.2
cachetools==4.2.2
certifi==2023.11.17
Django==5.0.1
python-telegram-bot==13.15
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tornado==6.1
tzlocal==5.2
